+++ src/Test/Integration/Model/Payment/FactoryTest.php
@@ -56,6 +56,7 @@ class FactoryTest extends TestCase
      * @throws LocalizedException
      *
      * @magentoConfigFixture current_store payment/braintree/active 1
+     * @magentoConfigFixture current_store payment/braintree_paypal/active 1
      */
     public function testPaymentMethodMapping($methodCode, $expectedPaymentMethod): void
     {
@@ -73,7 +74,8 @@ class FactoryTest extends TestCase
     {
         return [
             'return shopgate as default payment method' => ['PREPAY', 'shopgate'],
-            'Braintree Credit Card' => ['BRAINTR_CC', 'braintree']
+            'Braintree Credit Card'                     => ['BRAINTR_CC', 'braintree'],
+            'Braintree PayPal'                          => ['BRAINTR_PP', 'braintree_paypal']
         ];
     }
 }