+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- Mapping for Braintree PayPal payment method
+
 ## [2.9.13] - 2019-11-26
 ### Changed
 - Payment mapping is now based on factories